+++ ui/app/_components/cfa/CfaHeader.module.css
@@ -1,8 +1,3 @@
-.container {
-  margin: 0 auto;
-  padding: 0 1rem;
-}
-
 .title {
   color: var(--background-flat-blue-cumulus);
   margin-bottom: 2rem;
